# docs/openapi.yaml
basePath: /api/v1
definitions:
  apierror.Response:
    properties:
      code:
        type: string
        description: Stable machine-readable error code. Clients should branch on this field rather than
          message text.
      details:
        additionalProperties:
          type: string
        type: object
        description: Optional field-level validation details keyed by request field.
      message:
        type: string
        description: Safe human-readable error summary.
      request_id:
        type: string
        description: Server-generated ID for tracing this request; also returned in X-Request-ID.
    type: object
    description: Response representation used by this API.
    example:
      code: INVALID_REQUEST
      message: Invalid request
      request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      details:
        total: must be a decimal string
...
  model.CreateMatchRequest:
    properties:
      end_time:
        type: string
        description: RFC 3339 timestamp; must be later than start_time.
      start_time:
        type: string
        description: RFC 3339 timestamp; must be earlier than end_time.
      team_a:
        type: string
      team_b:
        type: string
      type:
        type: string
        description: Stable sport-type ID for this fixture.
    required:
    - end_time
    - start_time
    - team_a
    - team_b
    - type
    type: object
    description: CreateMatchRequest representation used by this API.
...
  model.MatchDto:
    properties:
      end_time:
        type: string
        description: Match end timestamp in RFC 3339 format.
      id:
        type: string
        description: Unique match identifier.
      is_draw:
        type: boolean
        description: True when the terminal result is a draw.
      start_time:
        type: string
        description: Match start timestamp in RFC 3339 format.
      team_a:
        type: string
      team_a_rate:
        example: '1.750000'
        type: string
        description: Current server-calculated rate for team_a, represented with six fractional digits.
      team_a_score:
        type: integer
      team_b:
        type: string
      team_b_rate:
        example: '2.500000'
        type: string
        description: Current server-calculated rate for team_b, represented with six fractional digits.
      team_b_score:
        type: integer
      type:
        type: string
      winner:
        type: string
        description: Winning team ID, or empty when no winner is set.
    type: object
    description: MatchDto representation used by this API.
...
  model.MatchesByDate:
    properties:
      date:
        type: string
        format: date-time
        description: UTC calendar date used to group the fixtures.
      types:
        items:
          $ref: '#/definitions/model.MatchesByType'
        type: array
    type: object
  model.MatchesByType:
    properties:
      matches:
        items:
          $ref: '#/definitions/model.MatchDto'
        type: array
      sportType:
        type: string
    type: object
...
  model.UpdateMatchRequest:
    properties:
      end_time:
        type: string
        description: Optional RFC 3339 timestamp.
      start_time:
        type: string
        description: Optional RFC 3339 timestamp.
      team_a:
        type: string
      team_b:
        type: string
      type:
        type: string
        description: Optional stable sport-type ID.
    type: object
    description: UpdateMatchRequest representation used by this API.
...
host: localhost:8080
info:
  title: Intania 888 API
  version: 1.0.0-breaking
  description: "Browser and application API for Intania 888. All application routes use the /api/v1 base\
    \ path. Browser authentication uses an HttpOnly session cookie; call GET /auth/me after login to retrieve\
    \ the current profile and session-bound csrf_token. Send that token in X-CSRF-Token with browser mutations\
    \ while a session is active and send an Origin allowed by the server. Mutations are rejected when\
    \ either check fails. \n\nMoney values are JSON strings. Responses use exactly two fractional digits;\
    \ requests accept up to two. Rate values are JSON strings with up to six fractional digits in requests\
    \ and six in responses. Numeric JSON tokens are not accepted for these decimal values. \n\nAPI failures\
    \ use the shared apierror.Response shape: code, message, request_id, and optional field-level details.\
    \ The same request ID is returned in the X-Request-ID header. Dates and timestamps use the formats\
    \ documented by each field. The separate /external/* backend integration is described in docs/README.md\
    \ and is intentionally outside this browser-focused reference.\n\nThe API is rate limited to 200 requests\
    \ per minute per client IP. Exceeding the limit returns 429 TOO_MANY_REQUESTS."
paths:
...
  /matches:
    get:
      parameters:
      - description: Optional sport-type ID used to filter fixtures.
        in: query
        name: typeId
        type: string
      - description: Select schedule or result grouping.
        in: query
        name: schedule
        type: string
        enum:
        - schedule
        - result
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MatchesByDate'
            type: array
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'Bad Request Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '500':
          description: 'Internal Server Error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: List matches
      tags:
      - Match
      description: Lists fixtures grouped by start date and sport type. Optionally filter by sport-type
        ID and choose schedule or result grouping.
    post:
      consumes:
      - application/json
      parameters:
      - description: Match
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/model.CreateMatchRequest'
      - name: X-CSRF-Token
        in: header
        description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        required: true
        type: string
      produces:
      - application/json
      responses:
        '201':
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'Bad Request Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'Forbidden Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '500':
          description: 'Internal Server Error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Create a match
      tags:
      - Match
      description: Administrator-only. Creates a fixture with two team IDs, a sport-type ID, and start/end
        timestamps. The end time must be later than the start time.
  /matches/{id}:
    delete:
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - name: X-CSRF-Token
        in: header
        description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        required: true
        type: string
      responses:
        '200':
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'Forbidden Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'Not Found Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'Internal Server Error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Delete a match
      tags:
      - Match
      description: Administrator-only. Deletes the selected fixture when domain constraints allow it.
        A missing match returns 404.
    get:
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/model.MatchDto'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              id: 1d93849c-a61c-4eb1-bcf4-4d27a44002bb
              team_a: team-a-id
              team_b: team-b-id
              type: BADMINTON_ALL
              start_time: '2026-10-01T09:00:00Z'
              end_time: '2026-10-01T10:00:00Z'
              team_a_rate: '1.750000'
              team_b_rate: '2.500000'
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'Not Found Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'Internal Server Error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Get a match by ID
      tags:
      - Match
      description: Returns one fixture with its schedule, scores, outcome, and current server-calculated
        rates for both teams.
    put:
      consumes:
      - application/json
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - description: Match fields to update
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/model.UpdateMatchRequest'
      - name: X-CSRF-Token
        in: header
        description: Session-bound token returned by GET /auth/me. Required for browser mutations.
        required: true
        type: string
      responses:
        '200':
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '400':
          description: 'Bad Request Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INVALID_REQUEST
              message: Invalid request
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '401':
          description: 'Unauthorized Failure body follows apierror.Response; code is stable, message is
            safe to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '403':
          description: 'Forbidden Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: FORBIDDEN
              message: The request is not allowed
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '404':
          description: 'Not Found Failure body follows apierror.Response; code is stable, message is safe
            to display, and request_id matches the X-Request-ID response header. Failure body follows
            apierror.Response: stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
        '500':
          description: 'Internal Server Error Failure body follows apierror.Response; code is stable,
            message is safe to display, and request_id matches the X-Request-ID response header. Failure
            body follows apierror.Response: stable code, safe message, request_id, and optional field
            details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: INTERNAL_ERROR
              message: Internal server error
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Update match details
      tags:
      - Match
      description: Administrator-only. Partially updates fixture details. Include at least one supported
        field; omitted fields remain unchanged. If both timestamps are present, end_time must be later
        than start_time.
...
  /matches/current/time:
    get:
      produces:
      - application/json
      responses:
        '200':
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
        '401':
          description: 'The browser session is missing or expired. Failure body follows apierror.Response:
            stable code, safe message, request_id, and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: UNAUTHORIZED
              message: Authentication required
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '503':
          description: 'Session or access-policy state could not be checked because a required dependency
            is unavailable. Failure body follows apierror.Response: stable code, safe message, request_id,
            and optional field details.'
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID: *id001
          examples:
            application/json:
              code: DEPENDENCY_UNAVAILABLE
              message: A required dependency is unavailable
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
        '429':
          description: The client IP exceeded the API limit of 200 requests per minute.
          schema:
            $ref: '#/definitions/apierror.Response'
          headers:
            X-Request-ID:
              type: string
              description: Correlation ID for this request. The same value appears in apierror.Response.request_id.
          examples:
            application/json:
              code: TOO_MANY_REQUESTS
              message: Too many requests
              request_id: 5ef9fba0-0123-4abc-8123-0123456789ab
      security:
      - CookieSession: []
      summary: Get current server match time
      tags:
      - Match
      description: Returns the backend’s current UTC time for clients that need to compare or display
        match schedules.
...
schemes:
- http
securityDefinitions:
  CookieSession:
    description: Browser session cookie. The cookie is HttpOnly and is normally sent by the browser automatically.
      Production uses the __Host-session cookie name. Swagger 2.0 represents this as a Cookie header.
      For browser mutations, also send X-CSRF-Token from GET /auth/me and an allowed Origin.
    in: header
    name: Cookie
    type: apiKey
swagger: '2.0'
consumes:
- application/json
produces:
- application/json
...